pyproject.toml
[project]
name = "tokreate"
version = "0.1.0"
description = "Paper QA backend"
authors = [
    {name = "Luca Soldaini", email = "lucas@allenai.org" }
]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyllms @ git+https://github.com/kagisearch/pyllms@58da611",
    "pysbd==0.3.4",
    "jinja2",
    "boto3",
    "numpy",
    "watchtower",
    "logging_json==0.4.0",
]

[project.urls]
"Homepage" = "https://github.com/soldni/tokreate"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"tokreate" = ["py.typed", "*.pyi"]

[build-system]
build-backend = "setuptools.build_meta"
requires = [
    "setuptools >= 61.0.0",
    "wheel"
]

[project.optional-dependencies]
dev = [
    "black>=22.6.0",
    "isort>=5.10.1",
    "mypy>=0.971",
    "pytest>=5.2",
    "ipython>=8.4.0",
    "autopep8>=1.7.0",
    "flake8>=5.0",
    "ipdb>=0.13.0",
    "flake8-pyi>=22.8.1",
    "Flake8-pyproject>=1.1.0"
]

[tool.black]
line-length = 115
include = '\.pyi?$'
exclude = '''
(
      __pycache__
    | \.git
    | \.mypy_cache
    | \.pytest_cache
    | \.vscode
    | \.venv
    | \bdist\b
    | \bdoc\b
)
'''

[tool.isort]
profile = "black"
line_length = 115
multi_line_output = 3

[tool.autopep8]
max_line_length = 115
in-place = true
recursive = true
aggressive = 3

[tool.mypy]
python_version = 3.8
ignore_missing_imports = true
no_site_packages = true
allow_redefinition = false
warn_unused_configs = true
warn_unused_ignores = true
warn_no_return = true
warn_return_any = false
warn_unreachable = true
show_error_codes = true
pretty = true

[tool.mypy-tests]
strict_optional = false

[tool.flake8]
per-file-ignores = [
    '__init__.py:F401',
    '*.pyi:E302,E305',
    '*.py:E203'
]
